I can fly

+++ graph_completion/CrossAdjacencyMatrix.py
@@ -7,10 +7,6 @@ import scipy.sparse as sp
 from scipy.optimize import linear_sum_assignment
 from graph_completion.CrossGraphCompletion import CrossGraphCompletion
 from graph_completion.functions import RelationWeighting
-from tools.print_time_info import print_time_info
-from tools.timeit import timeit
-from pprint import pprint
-
 
 def normalize_adj(adj):
     """Symmetrically normalize adjacency matrix."""
@@ -37,7 +33,7 @@ def g_func_template(a, b, c, d, e):
 
 def get_sparse_unit_matrix(size):
     values = torch.from_numpy(np.ones([size], dtype=np.float32))
-    poses = torch.from_numpy(np.asarray([[i for i in range(size)] for _ in range(2)]))
+    poses = torch.from_numpy(np.asarray([[i for i in range(size)] for _ in range(2)], dtype=np.int64))
     return torch_trans2sp(poses, values, (size, size))
 
 class CrossAdjacencyMatrix(nn.Module):
@@ -169,7 +165,7 @@ def build_adms_rconf_imp_pca(triples, new_triple_confs, num_entity, relation2con
             sp_matrix[0][pos] = 1
             sp_matrix[2][pos] = 1
     for key, sp_m in sp_matrix.items():
-        poses = torch.from_numpy(np.asarray(list(zip(*sp_m.keys())), dtype=np.int))
+        poses = torch.from_numpy(np.asarray(list(zip(*sp_m.keys())), dtype=np.int64))
         values = torch.from_numpy(np.asarray(list(sp_m.values()), dtype=np.float32))
         assert len(values) == len(poses[0]) == len(poses[-1])
         sp_matrix[key] = torch_trans2sp(poses, values, [num_entity, num_entity])
